build: implement flexible site-selection system
Implement a flexible system for handling site-defined features as well
as packages.

This system is inspired by the existing feature-system and allows for a
more flexible approach for selecting specific packages for devices.

Features are now defined in a `features` file in the site-root. The same
goes for packages.

These files are sequentially evaluated and the device-package list is
evaluated for each device independently.

Signed-off-by: David Bauer <mail@david-bauer.net>