build: allow to disable ALL_NONSHARED
When building in a CI where it is not intended to publish nor use
target-specific nonshared packaged this step increases the build-time.

Add a option GLUON_BUILD_NONSHARED which can optionally be disabled by
setting it to 0. This way, Gluon skips building these target-specific
packages if they are not required for the default package-set.

It is worth t note that this not only disables these packages. It can
also disable certain kernel config-options. Thus, the resulting image
might differ dependening on the state of this setting.

Signed-off-by: David Bauer <mail@david-bauer.net>